requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
auth==0.5.3
authy==2.2.6
blinker==1.4
cachetools==4.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.3.1
cryptography==3.3.1
defusedxml==0.6.0
Deprecated==1.2.10
diff-match-patch==20200713
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-3-jet==1.0.8
django-cors-headers==3.6.0
django-email-verification==0.0.2
django-environ==0.4.5
django-excel-data-sync==0.3.0
django-extensions==3.1.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-jet==1.0.8
django-phone-verify==2.0.0
django-phonenumber-field==5.0.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-csv==2.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
docutils==0.16
drf-renderer-xlsx==0.3.9
drf-yasg==1.20.0
et-xmlfile==1.0.1
eventlet==0.30.0
Faker==5.3.0
falcon==2.0.0
filelock==3.0.12
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
greenlet==0.4.17
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
lml==0.1.0
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
mongoengine==0.22.1
nexmo==2.5.2
numpy==1.19.0
oauthlib==3.1.0
odfpy==1.4.1
opencv-python==4.2.0.34
openpyxl==3.0.5
packaging==20.8
pandas==1.2.0
phonenumbers==8.12.15
Pillow==7.1.2
pluggy==0.13.1
protobuf==3.13.0
psycopg2==2.8.6
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-xlsx==0.6.0
PyJWT==1.7.1
pymongo==3.11.2
pyotp==2.4.1
pyparsing==2.4.7
pytest==6.2.1
pytest-cov==2.10.1
pytest-django==4.1.0
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejwt==2.0.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
texttable==1.6.3
toml==0.10.2
twilio==6.50.1
typing==3.7.4.3
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.11
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
xlrd==2.0.1
XlsxWriter==0.9.6
xlwt==1.3.0
